backend-M3M/requirements.txt
langchain
langchain-community
langchain-groq
langgraph
chromadb
sentence-transformers
python-dotenv
pandas
uvicorn
fastapi
pydantic
jinja2
